--- backend/src/app/auth-code/auth-code.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import {TypeOrmModule} from "@nestjs/typeorm";
-import {AuthCodeService} from "./auth-code.service";
-import {AuthCodeEntity} from "../../common/entities/auth-code.entity";
-
-@Module({
-    imports: [TypeOrmModule.forFeature([AuthCodeEntity])],
-    providers: [AuthCodeService],
-    controllers: [],
-    exports: [AuthCodeService]
-})
-export class AuthCodeModule {}
