--- backend/src/app/ingredients/ingredients.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import {TypeOrmModule} from "@nestjs/typeorm";
-import {IngredientsService} from "./ingredients.service";
-import {IngredientsEntity} from "../../common/entities/ingredients.entity";
-
-@Module({
-    imports: [TypeOrmModule.forFeature([IngredientsEntity])],
-    providers: [IngredientsService],
-    controllers: [],
-    exports: [IngredientsService]
-})
-export class IngredientsModule {}
