--- backend/src/app/kitchens/kitchens.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import {TypeOrmModule} from "@nestjs/typeorm";
-import {KitchensService} from "./kitchens.service";
-import {KitchensEntity} from "../../common/entities/kitchen.entity";
-
-@Module({
-    imports: [TypeOrmModule.forFeature([KitchensEntity])],
-    providers: [KitchensService],
-    controllers: [],
-    exports: [KitchensService]
-})
-export class KitchensModule {}
