--- backend/src/app/recipe-events/recipe-events.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
-import {RecipeEventsEntity} from "../../common/entities/recipe-events.entity";
-
-@Injectable()
-export class RecipeEventsService {
-    constructor(
-        @InjectRepository(RecipeEventsEntity)
-        private recipeEventsRepository: Repository<RecipeEventsEntity>,
-    ) {}
-}
