--- backend/src/app/recipes/recipes.service.ts
@@ -1,12 +1,0 @@
-import {RecipesEntity} from "../../common/entities/recipes.entity";
-import {Injectable} from "@nestjs/common";
-import {InjectRepository} from "@nestjs/typeorm";
-import {Repository} from "typeorm";
-
-@Injectable()
-export class RecipesService {
-    constructor(
-        @InjectRepository(RecipesEntity)
-        private recipesRepository: Repository<RecipesEntity>,
-    ) {}
-}
