--- backend/src/app/subscriptions-kitchen/subscriptions-kitchen.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import {TypeOrmModule} from "@nestjs/typeorm";
-import {SubscriptionsKitchenService} from "./subscriptions-kitchen.service";
-import {SubscriptionsKitchenEntity} from "../../common/entities/subscriptions-kitchen.entity";
-
-@Module({
-    imports: [TypeOrmModule.forFeature([SubscriptionsKitchenEntity])],
-    providers: [SubscriptionsKitchenService],
-    controllers: [],
-    exports: [SubscriptionsKitchenService]
-})
-export class SubscriptionsKitchenModule {}
