--- backend/src/app/subscriptions-users/subscriptions-users.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
-import {SubscriptionsUsersEntity} from "../../common/entities/subscriptions-users.entity";
-
-@Injectable()
-export class SubscriptionsUsersService {
-    constructor(
-        @InjectRepository(SubscriptionsUsersEntity)
-        private subscriptionsUserRepository: Repository<SubscriptionsUsersEntity>,
-    ) {}
-}
