--- backend/src/app/user_notification_setting/subscriptions-users.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
-import {UserNotificationSettingEntity} from "../../common/entities/user-notification-setting.entity";
-
-@Injectable()
-export class SubscriptionsUsersService {
-    constructor(
-        @InjectRepository(UserNotificationSettingEntity)
-        private userNotificationRepository: Repository<UserNotificationSettingEntity>,
-    ) {}
-}
