--- backend/src/app/users/users.service.ts
@@ -1,12 +1,0 @@
-import { Injectable } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-import { Repository } from "typeorm";
-import { UsersEntity } from "../../common/entities/users.entity";
-
-@Injectable()
-export class UsersService {
-    constructor(
-        @InjectRepository(UsersEntity)
-        private usersRepository: Repository<UsersEntity>,
-    ) {}
-}
