+++ backend/src/common/repository-services/auth-code/auth-code.service.ts
@@ -1,0 +1,12 @@
+import {Injectable} from "@nestjs/common";
+import {InjectRepository} from "@nestjs/typeorm";
+import {Repository} from "typeorm";
+import {AuthCodeEntity} from "../../entities/auth-code.entity";
+
+@Injectable()
+export class AuthCodeService {
+    constructor(
+        @InjectRepository(AuthCodeEntity)
+        private authCodeRepository: Repository<AuthCodeEntity>,
+    ) {}
+}
