+++ backend/src/common/repository-services/ingredients/ingredients.service.ts
@@ -1,0 +1,12 @@
+import {Injectable} from "@nestjs/common";
+import {InjectRepository} from "@nestjs/typeorm";
+import {Repository} from "typeorm";
+import {IngredientsEntity} from "../../entities/ingredients.entity";
+
+@Injectable()
+export class IngredientsService {
+    constructor(
+        @InjectRepository(IngredientsEntity)
+        private ingredientsRepository: Repository<IngredientsEntity>,
+    ) {}
+}
