+++ backend/src/common/repository-services/kitchens/kitchens.service.ts
@@ -1,0 +1,12 @@
+import {Injectable} from "@nestjs/common";
+import {InjectRepository} from "@nestjs/typeorm";
+import {Repository} from "typeorm";
+import {KitchensEntity} from "../../entities/kitchen.entity";
+
+@Injectable()
+export class KitchensService {
+    constructor(
+        @InjectRepository(KitchensEntity)
+        private kitchensRepository: Repository<KitchensEntity>,
+    ) {}
+}
