+++ backend/src/common/repository-services/portal-message/portal-message.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {PortalMessageService} from "./portal-message.service";
+import {PortalMessageEntity} from "../../entities/portal-message.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([PortalMessageEntity])],
+    providers: [PortalMessageService],
+    controllers: [],
+    exports: [PortalMessageService]
+})
+export class PortalMessageModule {}
