+++ backend/src/common/repository-services/portal-message/portal-message.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {PortalMessageEntity} from "../../entities/portal-message.entity";
+
+@Injectable()
+export class PortalMessageService {
+    constructor(
+        @InjectRepository(PortalMessageEntity)
+        private portalMessageRepository: Repository<PortalMessageEntity>,
+    ) {}
+}
