+++ backend/src/common/repository-services/products/products.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {ProductsService} from "./products.service";
+import {ProductsEntity} from "../../entities/products.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([ProductsEntity])],
+    providers: [ProductsService],
+    controllers: [],
+    exports: [ProductsService]
+})
+export class ProductsModule {}
