+++ backend/src/common/repository-services/products/products.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {ProductsEntity} from "../../entities/products.entity";
+
+@Injectable()
+export class ProductsService {
+    constructor(
+        @InjectRepository(ProductsEntity)
+        private productsRepository: Repository<ProductsEntity>,
+    ) {}
+}
