+++ backend/src/common/repository-services/recipe-events/recipe-events.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {RecipeEventsService} from "./recipe-events.service";
+import {RecipeEventsEntity} from "../../entities/recipe-events.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([RecipeEventsEntity])],
+    providers: [RecipeEventsService],
+    controllers: [],
+    exports: [RecipeEventsService]
+})
+export class RecipeEventsModule {}
