+++ backend/src/common/repository-services/recipe-events/recipe-events.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {RecipeEventsEntity} from "../../entities/recipe-events.entity";
+
+@Injectable()
+export class RecipeEventsService {
+    constructor(
+        @InjectRepository(RecipeEventsEntity)
+        private recipeEventsRepository: Repository<RecipeEventsEntity>,
+    ) {}
+}
