+++ backend/src/common/repository-services/recipes/recipes.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {RecipesService} from "./recipes.service";
+import {RecipesEntity} from "../../entities/recipes.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([RecipesEntity])],
+    providers: [RecipesService],
+    controllers: [],
+    exports: [RecipesService]
+})
+export class RecipesModule {}
