+++ backend/src/common/repository-services/refresh-tokens/refresh-tokens.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {RefreshTokensService} from "./refresh-tokens.service";
+import {RefreshTokensEntity} from "../../entities/refresh-tokens.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([RefreshTokensEntity])],
+    providers: [RefreshTokensService],
+    controllers: [],
+    exports: [RefreshTokensService]
+})
+export class RefreshTokensModule {}
