+++ backend/src/common/repository-services/refresh-tokens/refresh-tokens.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {RefreshTokensEntity} from "../../entities/refresh-tokens.entity";
+
+@Injectable()
+export class RefreshTokensService {
+    constructor(
+        @InjectRepository(RefreshTokensEntity)
+        private refreshTokensRepository: Repository<RefreshTokensEntity>,
+    ) {}
+}
