+++ backend/src/common/repository-services/subscriptions-kitchen/subscriptions-kitchen.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {SubscriptionsKitchenEntity} from "../../entities/subscriptions-kitchen.entity";
+
+@Injectable()
+export class SubscriptionsKitchenService {
+    constructor(
+        @InjectRepository(SubscriptionsKitchenEntity)
+        private subscriptionsKitchenRepository: Repository<SubscriptionsKitchenEntity>,
+    ) {}
+}
