+++ backend/src/common/repository-services/subscriptions-users/subscriptions-users.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {SubscriptionsUsersService} from "./subscriptions-users.service";
+import {SubscriptionsUsersEntity} from "../../entities/subscriptions-users.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([SubscriptionsUsersEntity])],
+    providers: [SubscriptionsUsersService],
+    controllers: [],
+    exports: [SubscriptionsUsersService]
+})
+export class SubscriptionsUsersModule {}
