+++ backend/src/common/repository-services/subscriptions-users/subscriptions-users.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {SubscriptionsUsersEntity} from "../../entities/subscriptions-users.entity";
+
+@Injectable()
+export class SubscriptionsUsersService {
+    constructor(
+        @InjectRepository(SubscriptionsUsersEntity)
+        private subscriptionsUserRepository: Repository<SubscriptionsUsersEntity>,
+    ) {}
+}
