+++ backend/src/common/repository-services/user-notification-setting/subscriptions-users.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {SubscriptionsUsersService} from "./subscriptions-users.service";
+import {UserNotificationSettingEntity} from "../../entities/user-notification-setting.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([UserNotificationSettingEntity])],
+    providers: [SubscriptionsUsersService],
+    controllers: [],
+    exports: [SubscriptionsUsersService]
+})
+export class SubscriptionsUsersModule {}
