+++ backend/src/common/repository-services/user-notification-setting/subscriptions-users.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import {UserNotificationSettingEntity} from "../../entities/user-notification-setting.entity";
+
+@Injectable()
+export class SubscriptionsUsersService {
+    constructor(
+        @InjectRepository(UserNotificationSettingEntity)
+        private userNotificationRepository: Repository<UserNotificationSettingEntity>,
+    ) {}
+}
