+++ backend/src/common/repository-services/users/users.module.ts
@@ -1,0 +1,12 @@
+import { Module } from '@nestjs/common';
+import {TypeOrmModule} from "@nestjs/typeorm";
+import {UsersService} from "./users.service";
+import {UsersEntity} from "../../entities/users.entity";
+
+@Module({
+    imports: [TypeOrmModule.forFeature([UsersEntity])],
+    providers: [UsersService],
+    controllers: [],
+    exports: [UsersService]
+})
+export class UsersModule {}
