+++ backend/src/common/repository-services/users/users.service.ts
@@ -1,0 +1,12 @@
+import { Injectable } from "@nestjs/common";
+import { InjectRepository } from "@nestjs/typeorm";
+import { Repository } from "typeorm";
+import { UsersEntity } from "../../entities/users.entity";
+
+@Injectable()
+export class UsersService {
+    constructor(
+        @InjectRepository(UsersEntity)
+        private usersRepository: Repository<UsersEntity>,
+    ) {}
+}
